+++ superset/assets/visualizations/deckgl/layers/screengrid.jsx
@@ -9,7 +9,7 @@ import { ScreenGridLayer } from 'deck.gl';
 import AnimatableDeckGLContainer from '../AnimatableDeckGLContainer';
 
 import * as common from './common';
-import { parseTimeGrain } from '../../../javascripts/modules/time';
+import { getPlaySliderParams } from '../../../javascripts/modules/time';
 import sandboxedEval from '../../../javascripts/modules/sandbox';
 
 function getPoints(data) {
@@ -62,20 +62,10 @@ class DeckGLScreenGrid extends React.PureComponent {
   /* eslint-disable no-unused-vars */
   static getDerivedStateFromProps(nextProps, prevState) {
     const fd = nextProps.slice.formData;
+
     const timeGrain = fd.time_grain_sqla || fd.granularity || 'minute';
-
-    // find start and end based on the data
     const timestamps = nextProps.payload.data.features.map(f => f.__timestamp);
-    let start = Math.min(...timestamps);
-    let end = Math.max(...timestamps);
-
-    // lock start and end to the closest steps
-    const step = parseTimeGrain(timeGrain);
-    start -= start % step;
-    end += step - end % step;
-
-    const values = timeGrain != null ? [start, start + step] : [start, end];
-    const disabled = timestamps.every(timestamp => timestamp === null);
+    const { start, end, step, values, disabled } = getPlaySliderParams(timestamps, timeGrain);
 
     return { start, end, step, values, disabled };
   }
